Move logger to common folder
